Use local local copy of db context in query
Not totally sure but.. I THINK uses the middleware to get a dbcontext from the pool which is presented as a local parameter. I think also that now we are returning this as a task to be executed in it's own thread and tasks need to return objects like lists instead of IQueryable? Maybe it becomes clear later.